requirements.txt
Django
djangorestframework
django-environ
mysqlclient